More warnings.

+++ code/ui/ui_saber.cpp
@@ -308,7 +308,7 @@ void UI_SaberLoadParms( void )
 	holdChar = saberExtensionListBuf;
 	for ( i = 0; i < fileCnt; i++, holdChar += saberExtFNLen + 1 )
 	{
-		saberExtFNLen = strlen( holdChar );
+		saberExtFNLen = static_cast<int>(strlen( holdChar ));
 
 		len = ui.FS_ReadFile( va( "ext_data/sabers/%s", holdChar), (void **) &buffer );
